requirements.txt
streamlit
scikit-learn
pandas
Pillow
numpy
joblib